backend/requirement.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cv==1.0.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.8
flatbuffers==25.1.24
fonttools==4.56.0
fpdf==1.7.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
groq==0.18.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.1.0
narwhals==1.25.2
nest-asyncio==1.6.0
numba==0.61.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.1.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
pywin32==308
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requirement.txt==0.1.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.3.0
scipy==1.15.1
seaborn==0.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.0
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
